+++ mybatis-plus-generator/src/main/java/com/baomidou/mybatisplus/generator/config/builder/Repository.java
@@ -29,51 +29,51 @@ import java.util.HashMap;
 import java.util.Map;
 
 /**
- * Service属性配置
+ * Repository属性配置
  *
  * @author nieqiurong 2020/10/11.
  * @since 3.5.0
  */
-public class Service implements ITemplate {
+public class Repository implements ITemplate {
 
-    private final static Logger LOGGER = LoggerFactory.getLogger(Service.class);
+    private final static Logger LOGGER = LoggerFactory.getLogger(Repository.class);
 
-    private Service() {
+    private Repository() {
     }
 
     /**
-     * 自定义继承的Service类全称，带包名
+     * 自定义继承的Repository类全称，带包名
      */
-    private String superServiceClass = ConstVal.SUPER_SERVICE_CLASS;
+    private String superRepositoryClass = ConstVal.SUPER_REPOSITORY_CLASS;
 
     /**
-     * 自定义继承的ServiceImpl类全称，带包名
+     * 自定义继承的RepositoryImpl类全称，带包名
      */
-    private String superServiceImplClass = ConstVal.SUPER_SERVICE_IMPL_CLASS;
+    private String superRepositoryImplClass = ConstVal.SUPER_REPOSITORY_IMPL_CLASS;
 
     @NotNull
-    public String getSuperServiceClass() {
-        return superServiceClass;
+    public String getSuperRepositoryClass() {
+        return superRepositoryClass;
     }
 
     @NotNull
-    public String getSuperServiceImplClass() {
-        return superServiceImplClass;
+    public String getSuperRepositoryImplClass() {
+        return superRepositoryImplClass;
     }
 
     /**
-     * 转换输出Service文件名称
+     * 转换输出Repository文件名称
      *
      * @since 3.5.0
      */
-    private ConverterFileName converterServiceFileName = (entityName -> "I" + entityName + ConstVal.SERVICE);
+    private ConverterFileName converterRepositoryFileName = (entityName -> "I" + entityName + ConstVal.REPOSITORY);
 
     /**
-     * 转换输出ServiceImpl文件名称
+     * 转换输出RepositoryImpl文件名称
      *
      * @since 3.5.0
      */
-    private ConverterFileName converterServiceImplFileName = (entityName -> entityName + ConstVal.SERVICE_IMPL);
+    private ConverterFileName converterRepositoryImplFileName = (entityName -> entityName + ConstVal.REPOSITORY_IMPL);
 
     /**
      * 是否覆盖已有文件（默认 false）
@@ -83,13 +83,13 @@ public class Service implements ITemplate {
     private boolean fileOverride;
 
     @NotNull
-    public ConverterFileName getConverterServiceFileName() {
-        return converterServiceFileName;
+    public ConverterFileName getConverterRepositoryFileName() {
+        return converterRepositoryFileName;
     }
 
     @NotNull
-    public ConverterFileName getConverterServiceImplFileName() {
-        return converterServiceImplFileName;
+    public ConverterFileName getConverterRepositoryImplFileName() {
+        return converterRepositoryImplFileName;
     }
 
     public boolean isFileOverride() {
@@ -100,107 +100,107 @@ public class Service implements ITemplate {
     @NotNull
     public Map<String, Object> renderData(@NotNull TableInfo tableInfo) {
         Map<String, Object> data = new HashMap<>();
-        data.put("superServiceClassPackage", this.superServiceClass);
-        data.put("superServiceClass", ClassUtils.getSimpleName(this.superServiceClass));
-        data.put("superServiceImplClassPackage", this.superServiceImplClass);
-        data.put("superServiceImplClass", ClassUtils.getSimpleName(this.superServiceImplClass));
+        data.put("superRepositoryClassPackage", this.superRepositoryClass);
+        data.put("superRepositoryClass", ClassUtils.getSimpleName(this.superRepositoryClass));
+        data.put("superRepositoryImplClassPackage", this.superRepositoryImplClass);
+        data.put("superRepositoryImplClass", ClassUtils.getSimpleName(this.superRepositoryImplClass));
         return data;
     }
 
     public static class Builder extends BaseBuilder {
 
-        private final Service service = new Service();
+        private final Repository repository = new Repository();
 
         public Builder(@NotNull StrategyConfig strategyConfig) {
             super(strategyConfig);
         }
 
         /**
-         * Service接口父类
+         * Repository接口父类
          *
          * @param clazz 类
          * @return this
          */
-        public Builder superServiceClass(@NotNull Class<?> clazz) {
-            return superServiceClass(clazz.getName());
+        public Builder superRepositoryClass(@NotNull Class<?> clazz) {
+            return superRepositoryClass(clazz.getName());
         }
 
         /**
-         * Service接口父类
+         * Repository接口父类
          *
-         * @param superServiceClass 类名
+         * @param superRepositoryClass 类名
          * @return this
          */
-        public Builder superServiceClass(@NotNull String superServiceClass) {
-            this.service.superServiceClass = superServiceClass;
+        public Builder superRepositoryClass(@NotNull String superRepositoryClass) {
+            this.repository.superRepositoryClass = superRepositoryClass;
             return this;
         }
 
         /**
-         * Service实现类父类
+         * Repository实现类父类
          *
          * @param clazz 类
          * @return this
          */
-        public Builder superServiceImplClass(@NotNull Class<?> clazz) {
-            return superServiceImplClass(clazz.getName());
+        public Builder superRepositoryImplClass(@NotNull Class<?> clazz) {
+            return superRepositoryImplClass(clazz.getName());
         }
 
         /**
-         * Service实现类父类
+         * Repository实现类父类
          *
-         * @param superServiceImplClass 类名
+         * @param superRepositoryImplClass 类名
          * @return this
          */
-        public Builder superServiceImplClass(@NotNull String superServiceImplClass) {
-            this.service.superServiceImplClass = superServiceImplClass;
+        public Builder superRepositoryImplClass(@NotNull String superRepositoryImplClass) {
+            this.repository.superRepositoryImplClass = superRepositoryImplClass;
             return this;
         }
 
         /**
-         * 转换输出service接口文件名称
+         * 转换输出repository接口文件名称
          *
          * @param converter 　转换处理
          * @return this
          * @since 3.5.0
          */
-        public Builder convertServiceFileName(@NotNull ConverterFileName converter) {
-            this.service.converterServiceFileName = converter;
+        public Builder convertRepositoryFileName(@NotNull ConverterFileName converter) {
+            this.repository.converterRepositoryFileName = converter;
             return this;
         }
 
         /**
-         * 转换输出service实现类文件名称
+         * 转换输出repository实现类文件名称
          *
          * @param converter 　转换处理
          * @return this
          * @since 3.5.0
          */
-        public Builder convertServiceImplFileName(@NotNull ConverterFileName converter) {
-            this.service.converterServiceImplFileName = converter;
+        public Builder convertRepositoryImplFileName(@NotNull ConverterFileName converter) {
+            this.repository.converterRepositoryImplFileName = converter;
             return this;
         }
 
         /**
-         * 格式化service接口文件名称
+         * 格式化repository接口文件名称
          *
          * @param format 　格式
          * @return this
          * @since 3.5.0
          */
-        public Builder formatServiceFileName(@NotNull String format) {
-            return convertServiceFileName((entityName) -> String.format(format, entityName));
+        public Builder formatRepositoryFileName(@NotNull String format) {
+            return convertRepositoryFileName((entityName) -> String.format(format, entityName));
         }
 
         /**
-         * 格式化service实现类文件名称
+         * 格式化repository实现类文件名称
          *
          * @param format 　格式
          * @return this
          * @since 3.5.0
          */
-        public Builder formatServiceImplFileName(@NotNull String format) {
-            return convertServiceImplFileName((entityName) -> String.format(format, entityName));
+        public Builder formatRepositoryImplFileName(@NotNull String format) {
+            return convertRepositoryImplFileName((entityName) -> String.format(format, entityName));
         }
 
         /**
@@ -211,7 +211,7 @@ public class Service implements ITemplate {
         @Deprecated
         public Builder fileOverride() {
             LOGGER.warn("fileOverride方法后续会删除，替代方法为enableFileOverride方法");
-            this.service.fileOverride = true;
+            this.repository.fileOverride = true;
             return this;
         }
 
@@ -219,13 +219,13 @@ public class Service implements ITemplate {
          * 覆盖已有文件
          */
         public Builder enableFileOverride() {
-            this.service.fileOverride = true;
+            this.repository.fileOverride = true;
             return this;
         }
 
         @NotNull
-        public Service get() {
-            return this.service;
+        public Repository get() {
+            return this.repository;
         }
     }
 }